add header and cart styling. conditional modify options

+++ client/src/components/CartItem.js
@@ -2,44 +2,68 @@ import React from 'react'
 import styled from 'styled-components';
 import { removeFromCart, updateQuantity } from '../store/reducers/cartReducer';
 import { useDispatch, useSelector } from 'react-redux';
+import { Link, useLocation } from 'react-router-dom';
 
 function CartItem() {
+  const location = useLocation();
   const dispatch = useDispatch();
   const { items: cartItems } = useSelector(state => state.cart);
 
+  let modify = false;
+
+  if (location.pathname === '/cart') {
+    modify = true 
+  } else {
+    modify = false
+  }
+
   const handleRemoveFromCart = (itemId, size) => {
-    const item = cartItems.find((item) => item.product.id === itemId && item.size === size);
-    dispatch(removeFromCart({ product: item.product, size }));
+    const item = cartItems.find((item) => item.product.id === itemId && item.size === size); // find item in the cart that matches the given item ID and size
+    
+    if (item) { // safeguard against undefined items
+      dispatch(removeFromCart({ product: item.product, size })); // if item exists, remove from cart
+    }
   };
   
   const handleUpdateQuantity = (productId, size, newQuantity) => {
-    console.log('handleUpdateQuantity called with', productId, size, newQuantity);
+    const item = cartItems.find((item) => item.product.id === productId && item.size === size);
 
-    dispatch(updateQuantity({ productId, size, quantity: newQuantity }));
+    if (newQuantity > 0 && item) { 
+      dispatch(updateQuantity({ productId, size, quantity: newQuantity }));
+    } 
+    if (newQuantity === 0 && item) { 
+      dispatch(removeFromCart({ product: item.product, size }));
+    } 
   };
 
   return (
     <>
-      {cartItems.map((item, index) => (
-        <StyledCartItem key={`${item.product.id}-${index}`}>
-          <ItemImage>
-            <img src={item.product.images[0]} alt={item.product.model} />
-          </ItemImage>
+      {cartItems.map((item) => (
+         <StyledCartItem key={`${item.product.id}-${item.size}`}>
+          <Link to={`/${item.product.id}`}>
+            <ItemImage>
+              <img src={item.product.images[0]} alt={item.product.model} />
+            </ItemImage>
+          </Link>
           <ItemInfo>
             <ItemDetails>
-              <h3>{item.product.brand} {item.product.model}</h3>
+              <Link to={`/${item.product.id}`}><h3>{item.product.brand} {item.product.model}</h3></Link>
               <p>Size: {item.size}</p>
-              <p>Quantity: {item.quantity}</p>
-              <button onClick={() => handleRemoveFromCart(item.product.id, item.size)}>Remove</button>
+              { modify && <p>Quantity: {item.quantity}</p>}
+              { modify && 
               <Quantity>
-                <button onClick={() => handleUpdateQuantity(item.product.id, item.size, item.quantity - 1)}>-</button>
-                <button>{item.quantity}</button>
-                <button onClick={() => handleUpdateQuantity(item.product.id, item.size, item.quantity + 1)}>+</button>
+              <a onClick={() => handleUpdateQuantity(item.product.id, item.size, item.quantity - 1)}>-</a>
+              <input type="number" value={item.quantity} onChange={(e) => handleUpdateQuantity(item.product.id, item.size, parseInt(e.target.value))} />
+              <a onClick={() => handleUpdateQuantity(item.product.id, item.size, item.quantity + 1)}>+</a>
               </Quantity>
+              }
             </ItemDetails>
-            <ItemPrice>
-              <p>{item.product.price * item.quantity} kr.</p>
-            </ItemPrice>
+              <ItemPrice>
+                <p>{item.product.price * item.quantity} kr.</p>
+                { modify && 
+                <a onClick={() => handleRemoveFromCart(item.product.id, item.size)}>x</a>
+                }
+              </ItemPrice>
           </ItemInfo>
         </StyledCartItem>
       ))}
@@ -51,7 +75,7 @@ const StyledCartItem = styled.div`
 display: flex;
 flex: 1;
 width: 100%;
-margin-bottom: 1rem;
+margin-bottom: 1rem;  
 `
 const ItemImage = styled.div`
 width: 150px;
@@ -74,19 +98,47 @@ h3{
   font-weight: 500;
 }
 p{
-    color: var(--color-text)
+    color: var(--color-text);
   }
 `
 
 const ItemDetails = styled.div`
+display: flex;
+flex-direction: column;
+a:nth-child(4){
+  text-decoration: underline;
+}
 `
 const Quantity = styled.div`
+margin-top: 1.5rem;
 display: flex;
+justify-content: flex-start;
+margin-right: 2rem;
+input{
+  text-align: center;
+  height: 25px;
+  width: 40px;
+  border: 1px solid black;
+  font-size: 1rem;
+}
+a{
+  width: 20px;
+  height: 25px;
+  text-align: center;
+  color: white;
+  background: black;
+}
 `
 
 const ItemPrice = styled.div`
+display: flex;
+flex-direction: row;  
 p{
   color: black;
+}
+a{
+  color: var(--color-text);
+  margin-left: 1rem;
 }
 `
